Rename submodule under `cupy.prof` package
