+++ cupy/prof/__init__.py
@@ -1,2 +1,2 @@
-from cupy.prof.time_range import time_range  # NOQA
-from cupy.prof.time_range import TimeRangeDecorator  # NOQA
+from cupy.prof._time_range import time_range  # NOQA
+from cupy.prof._time_range import TimeRangeDecorator  # NOQA
