+++ cupy/prof/_time_range.py
